checks for notes for access only to the authenticated users

+++ backend/src/controllers/notes.ts
@@ -2,11 +2,15 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 // infers the type of the arguments in getNotes
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
+
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authenticatedUserId);
+    const notes = await NoteModel.find({ userId: authenticatedUserId }).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -16,8 +20,10 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 // Singular note by Id
 export const getNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid Note Id"); //Not up to the Id syntax standard
     }
@@ -25,6 +31,11 @@ export const getNote: RequestHandler = async (req, res, next) => {
     if (!note) {
       throw createHttpError(404, "Note not found"); //Not a note that exists in the DB with correct Id syntax
     }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
+    }
+
     res.status(200).json(note);
   } catch (error) {
     next(error);
@@ -45,13 +56,17 @@ export const createNote: RequestHandler<
 > = async (req, res, next) => {
   const title = req.body.title;
   const text = req.body.text;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!title) {
       throw createHttpError(400, "Note requires a title");
     }
 
     const newNote = await NoteModel.create({
+      userId: authenticatedUserId,
       title: title,
       text: text,
     });
@@ -81,9 +96,10 @@ export const updateNote: RequestHandler<
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newText = req.body.text;
+  const authenticatedUserId = req.session.userId;
 
   try {
-    //Error checking
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid Note Id");
     }
@@ -96,6 +112,9 @@ export const updateNote: RequestHandler<
     const note = await NoteModel.findById(noteId).exec();
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
     }
     note.title = newTitle;
     note.text = newText;
@@ -110,14 +129,19 @@ export const updateNote: RequestHandler<
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid Note Id");
     }
     const note = await NoteModel.findById(noteId).exec();
     if (!note) {
       throw createHttpError(404, "Note could not be found");
+    }
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
     }
     await note.deleteOne();
     res.sendStatus(204);
